+++ src/mkvtoolnix-gui/chapter_editor/tool.h
@@ -53,7 +53,6 @@ public slots:
   virtual void reload();
   virtual void openFiles(QStringList const &fileNames);
   virtual void openFilesFromCommandLine(QStringList const &fileNames);
-  virtual void setupTabPositions();
 
   virtual void removeChaptersFromExistingMatroskaFile();