+++ src/mkvtoolnix-gui/merge/tool.h
@@ -74,8 +74,6 @@ public slots:
   virtual void openMultipleConfigFilesFromCommandLine(QStringList const &fileNames);
   virtual void addMultipleFilesToNewSettings(QStringList const &fileNames, bool newSettingsForEachFile);
 
-  virtual void setupTabPositions();
-
 protected:
   Tab *appendTab(Tab *tab);
   virtual Tab *currentTab();