+++ src/mkvtoolnix-gui/watch_jobs/tool.h
@@ -51,7 +51,6 @@ public slots:
   virtual void saveCurrentTabOutput();
   virtual void saveAllTabs();
   virtual void enableMenuActions();
-  virtual void setupTabPositions();
   virtual void retranslateUi();
   virtual void switchToCurrentJobTab();